LRCI-3823 Rename package gitbranch to git.branch
