+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/git/branch/BaseGitBranchEntity.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.gitbranch;
+package com.liferay.jethr0.git.branch;
 
 import com.liferay.jethr0.entity.BaseEntity;
 import com.liferay.jethr0.job.JobEntity;
